# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "成果云系统API文档",
    "description": "成果云系统后端接口文档（Flask + MySQL）",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "登录后返回的 Bearer Token，格式：Bearer <access_token>"
    }
  },
...
  "paths": {
...
    "/api/students/": {
      "get": {
        "tags": ["学生"],
        "summary": "学生列表（分页/搜索）",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回学生列表"}}
      },
      "post": {
        "tags": ["学生"],
        "summary": "添加学生（教师录入直接生效）",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "body", "in": "body", "schema": {"type": "object"}}
        ],
        "responses": {"201": {"description": "添加成功"}}
      }
    },
    "/api/students/{id}": {
      "parameters": [
        {"name": "id", "in": "path", "required": true, "type": "integer"}
      ],
      "get": {
        "tags": ["学生"],
        "summary": "学生详情",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "返回学生详情"}}
      },
      "put": {
        "tags": ["学生"],
        "summary": "编辑学生",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "更新成功"}}
      },
      "delete": {
        "tags": ["学生"],
        "summary": "删除学生（成果保留）",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "删除成功"}}
      }
    },
    "/api/students/{id}/reset-password": {
      "post": {
        "tags": ["学生"],
        "summary": "重置学生密码",
        "security": [{"BearerAuth": []}],
        "responses": {"200": {"description": "重置成功"}}
      }
    },
...
  }
}